RED: Bind the name to an Angular scope value
We want to see the integration test turn red here, just to be sure that
Angular is running and picks up our binding. We do this before touching
any scope variables, so we know it turns green for the right reason.

If we changed the controller code before the view, we could end up
testing a false negative; if we misspelled `ng-bind` so Angular didn't
pick it up, we would still see the hard-coded view value - and we'd have
to troubleshoot this earlier change once the controller tries to expose
more than one value.