RED: Expect the controller to fetch its list of venues
It's necessary for this test to have a different expectation than the
previous, hard-coded venue.

If we had used "Loungen" in our expectation, we would have realized we
were missing `$httpBackend.flush()` - but once we implemented a GET
request in our controller, the test would then be green again -
without having to actually treat the returned data in any way.