RED: Expect the venue selection to be randomized
Note how the test controls the random values by mocking them. This is by
far the best way to ensure the application is acting on random, and
specifically on a library function.

An even better way to do this is to dependency-inject a random number
generator. It is, however, also massive overkill for a small application
like this.