GREEN: Allow for random selection in the integration test
This is a dead-simple solution which probably won't hold water for long,
but it's enough to support our simple demo app.

Notice how the expectation suddenly becomes asynchronous. We're dealing
with a leaky abstraction: all calls to the browser are done
asynchronously, but the `expect()` wrapper hides it by knowing about
promises and resolving them. Assertions like `.toContain()`, however, do
not have the same magic properties, so we must handle it ourselves.